v0.8 Sprint 1 (QA-012): decompose Slf4jTelemetrySink — close GodClass + TooManyMethods + CyclomaticComplexity (#124)
Continuation of v0.8 Sprint 1 PMD GodClass closure track (after QA-010
CommunityPersistenceEngine and QA-011 CommunityHttpRequestProcessor).

Slf4jTelemetrySink (429 LOC) carried a class-level @SuppressWarnings block
covering GodClass + TooManyMethods + CyclomaticComplexity + 5 others —
the "high WMC from JSON serialization; no clean decomposition without
allocation" rationale was outdated. JSON serialization can in fact be
split along stable internal seams without touching the no-allocation
contract (which only applies to core hot-path sinks; this is a
diagnostic/fallback sink and intentionally allocates).

Decomposition (3-way split along the JSON serialization concern):

- Slf4jTelemetryJsonEscaper — leaf string-escape helper. Owns
  escapeJson + appendJsonString + appendUnicodeEscape.

- Slf4jTelemetryRawArgsWriter — middle layer rawArgs JSON-array writer.
  Walks the disclosed rawArgs payload and serializes each element to JSON;
  delegates string escaping to JsonEscaper.

- Slf4jTelemetryJsonWriter — top entry. Composes the canonical six-field
  event line (timestamp/level/code/component/message/rawArgs); delegates
  rawArgs to RawArgsWriter and string escaping to JsonEscaper.

- Slf4jTelemetryLogLevelResolver — log-level routing extracted to its
  own helper. EX-prefix → LogLevel mapping uses a static prefix table
  instead of long boolean ladders (CC drops from 12 to 2 on mapCodeLevel).

Slf4jTelemetrySink retains the emit lifecycle (closed-flag guard, MDC
scope, log adapter routing), the no-op metric methods (increment/gauge/
latency contract preserved), and the test-facing adapter/scope
interfaces (LogAdapter / MdcAdapter / MdcScope + Slf4j* impls). Public
API surface unchanged.

Suppressions delta on the sink:
  removed: PMD.GodClass, PMD.TooManyMethods, PMD.CyclomaticComplexity,
           PMD.AvoidDuplicateLiterals, PMD.UseVarargs,
           PMD.ImplicitFunctionalInterface  (6 closed)
  retained: PMD.UseTryWithResources (MDC scope close order is explicit
            LIFO — try-with-resources cannot express),
            PMD.CloseResource (MdcScope is an AutoCloseable lambda —
            PMD doesn't track lambda close semantics)

MdcAdapter + MdcScope now annotated @FunctionalInterface (intent
clarification; they were already SAM types via lambda usage in
Slf4jMdcAdapter and pushMdc closure).

Test coverage: 17/17 tests green across Slf4jTelemetrySinkTckTest (8),
Slf4jTelemetrySinkDisclosureTest (3), Slf4jTelemetrySinkTest$ExMapping
(5), Slf4jTelemetrySinkTest$Lifecycle (1). No test-side changes — the
nested LogAdapter/MdcAdapter/MdcScope interfaces remain the test
extension surface and were preserved verbatim.

Full reactor verify SUCCESS: SPI + TCK + Core + Community + Community
Testkit + Build Config all green; PMD 0 violations on community
module; Checkstyle 0 violations on community module.
